requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
geopy==2.4.1
requests==2.31.0.0
folium==0.15.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
python-dotenv==1.0.0
alembic==1.12.1
pillow==11.0.0
aiofiles==23.2.1
pydantic-settings==2.1.0
jinja2==3.1.2
httpx==0.25.2
astral==3.2
geopy==2.4.1
psycopg2-binary==2.9.9
google-generativeai==0.3.2
